requirements.txt
pdfminer3k==1.3.1
pyquery==1.2.17
urllib3==1.7.1
reportlab==3.3.0